basic derivation runner
This is the start of a nix-inspired blog post generator with (hopefully)
very fast incremental rebuilds. I plan on writing the derivation build
plans in some sort of scripting language so that things like tags will
be simpler (import-from-derivation going to be a little tricky tho).
Also want to have it so that remote images are cached according to
timestamp instead of content hash. Overall a lot of things I _want_ but
going to be pretty hard to do them I think. Ah well that's why this is
just a hobby project